requirements.txt
python-telegram-bot==20.3
httpx
nest_asyncio
aiohttp==3.9.1
